Add non_spatial_dataset field
A spatial dataset is one that has geographic data associated with it, such as latitude and longitude coordinates or some form of location information. A non-spatial dataset, on the other hand, does not have geographic data associated with it.

By having a non_spatial_dataset metadata element in your schema, you can easily mark a dataset as non-spatial (if non_spatial_dataset is True) or spatial (if non_spatial_dataset is False). This can make it easier to find and filter records in your application, as you can quickly filter records based on whether they are spatial or not.